finish

+++ week12/矩阵加法模板/16-b2-2.hpp
@@ -7,26 +7,26 @@ private:
 	T value[R][C];
 public:
 	matrix() {
-		for (int i = 0; i < R; i++)
-			for (int j = 0; j < C; j++)
+		for (long unsigned int i = 0; i < R; i++)
+			for (long unsigned int j = 0; j < C; j++)
 				value[i][j] = T();
 	}
 	matrix<T, R, C> operator+(const matrix <T, R, C>& M)const {
 		matrix<T, R, C> temp;
-		for (int i = 0; i < R; i++)
-			for (int j = 0; j < C; j++)
+		for (long unsigned int i = 0; i < R; i++)
+			for (long unsigned int j = 0; j < C; j++)
 				temp.value[i][j] = value[i][j] + M.value[i][j];
 		return temp;
 	}
 	friend istream& operator >> (istream&in, matrix<T, R, C>&M) {
-		for (int i = 0; i < R; i++)
-			for (int j = 0; j < C; j++)
+		for (long unsigned int i = 0; i < R; i++)
+			for (long unsigned int j = 0; j < C; j++)
 				in >> M.value[i][j];
 		return in;
 	}
 	friend ostream& operator << (ostream&out, const matrix<T, R, C>&M) {
-		for (int i = 0; i < R; i++) {
-			for (int j = 0; j < C; j++)
+		for (long unsigned int i = 0; i < R; i++) {
+			for (long unsigned int j = 0; j < C; j++)
 				out << M.value[i][j]<<" ";
 			out << endl;
 		}
